fix: turn return statement into boolean value
for example user has id 1 and in array save there is [2, 3, 1]. This will return [1] which is not a boolean value. but if we use [1].length -> 1 which is number (we want true). Same case of id 4, This will return [] so [].length -> 0 which is number as well (we want false). However to turn this into boolean by using !!: !1 - > false -> !false -> true, !0 - > true -> !true -> false.